FIX: move validationSchema into utils folder FIX: rename validationSchema to ts
